backend/python/requirements.txt
fastapi
uvicorn
sentence-transformers
torch
requests
python-dotenv